trace: stop disabling top-level definitions in instrument-module
Instead, this will be handled by the tracing language by using
define/ctc-helper forms and the definitions for these forms introduced
in the flow-trace repository commit 632f82b.
This works because the require form of the flow-trace language ignores
requires of the ctc-helper modules, so they will not shadow the lang's
versions of those define/ctc-helper forms.