Change snake harness to use for/fold instead of pathological recursion
The original implementation uses a recursive loop using named let, but
this requires traversing the entire list on every iteration, since car
and cdr apply a label when they are applied. Since the list is almost
50,000 elements long, this is unworkable. Fortunately, a for/fold loop
is semantically equivalent, and it doesn’t affect the program trace in a
particularly meaningful way (though it does dramatically reduce the
trace’s size).