mutant-factory & mutant-runner: Avoid depending on (current-directory)
This change adjusts all references to files inside this repository to
use define-runtime-path instead of depending on a particular value for
(current-directory). This makes it possible to run mutant-factory and
mutant-runner from an arbitrary directory, even one outside this
repository. Output files will still be stored relative to the current
working directory of the runner process.