Tweak some contracts to avoid traced operations getting upset
The benchmarks use untraced versions of operations like `lambda`, but
recent changes to flow-trace use traced operations for primitives by
default, and they get unhappy if they receive a procedure that isn’t
saved as a base value. This mostly avoids that problem, but also makes a
couple similar tweaks.